Move mmap_* modules under src/mmap/
No functional changes, other than the module paths of mmap related
types.
