+++ coverage_config_aarch64.json
@@ -1,5 +1,5 @@
 {
   "coverage_score": 85.2,
-  "exclude_path": "mmap_windows.rs",
+  "exclude_path": "mmap/windows.rs",
   "crate_features": "backend-mmap,backend-atomic,backend-bitmap"
 }
