+++ src/lib.rs
@@ -53,15 +53,6 @@ pub use guest_memory::{
 pub mod io;
 pub use io::{ReadVolatile, WriteVolatile};
 
-#[cfg(all(feature = "backend-mmap", not(feature = "xen"), unix))]
-mod mmap_unix;
-
-#[cfg(all(feature = "backend-mmap", feature = "xen", unix))]
-mod mmap_xen;
-
-#[cfg(all(feature = "backend-mmap", windows))]
-mod mmap_windows;
-
 #[cfg(feature = "backend-mmap")]
 pub mod mmap;
